requirements.txt
asgiref==3.8.1
Django==5.2
sqlparse==0.5.3
typing_extensions==4.13.2

djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
python-dotenv==1.1.0
PyMuPDF==1.25.1
python-docx==1.0.0
openai==1.3.0
httpx==0.25.2
python-multipart==0.0.6

# Development
pytest==7.4.3
pytest-django==4.8.0
black==23.11.0
flake8==7.0.0
